chore: return null instead default
